ollama-ocr/api/requirements.txt
fastapi
uvicorn
requests
Pillow
opencv-python
pdf2image
PyMuPDF
jsonpickle